refactor: #196 launch cli tui
Add new package runner and `RunCLI`,`RunTUI`, remove `internal/tui/runner.go`
